pyproject.toml
[project]
name = "ai-agent-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
    "strands-agents>=1.28.0",
]
